Immediately stop worker when RSS exceeds max RSS
Relying on the sidekiq apis means getting sidekiq to stop once the memory
limit is exceeded involves:
- Spinning up a new thread
- Make a redis call
- Waiting for the sidekiq heartbeat thread to run

When servers are under heavy resource load this is too much and can result in RSS
continuing to grow and jobs continuing to process. So, immediately quiet the worker
to stop processing ASAP while we let everything else settle out.